+++ packages/plugin-defender-hardhat/test/propose-upgrade-uups.js
@@ -8,6 +8,7 @@ const { FormatTypes } = require('ethers/lib/utils');
 const { AdminClient } = require('defender-admin-client');
 
 const proposalUrl = 'https://example.com';
+const multisig = '0xc0889725c22e2e36c524F41AECfddF5650432464';
 
 test.beforeEach(async t => {
   t.context.fakeClient = sinon.createStubInstance(AdminClient);
@@ -32,13 +33,35 @@ test.afterEach.always(() => {
   sinon.restore();
 });
 
+test('proposes an upgrade and get tx response', async t => {
+  const { proposeUpgrade, fakeClient, greeter, GreeterV2 } = t.context;
+  fakeClient.proposeUpgrade.resolves({ url: proposalUrl });
+
+  const title = 'My upgrade';
+  const description = 'My contract upgrade';
+  const proposal = await proposeUpgrade(greeter.address, GreeterV2, { title, description, multisig });
+
+  t.is(proposal.url, proposalUrl);
+
+  t.not(proposal.txResponse.hash, undefined);
+  const txReceipt = await proposal.txResponse.wait();
+  t.not(txReceipt.contractAddress, undefined);
+
+  const proposal2 = await proposeUpgrade(greeter.address, GreeterV2, { title, description, multisig });
+
+  // even though impl was already deployed in first proposal, it should still provide a tx response for the same tx hash
+  t.is(proposal2.txResponse.hash, proposal.txResponse.hash);
+  const txReceipt2 = await proposal.txResponse.wait();
+  t.is(txReceipt2.contractAddress, txReceipt.contractAddress);
+});
+
 test('proposes an upgrade', async t => {
   const { proposeUpgrade, fakeClient, greeter, GreeterV2 } = t.context;
   fakeClient.proposeUpgrade.resolves({ url: proposalUrl });
 
   const title = 'My upgrade';
   const description = 'My contract upgrade';
-  const proposal = await proposeUpgrade(greeter.address, GreeterV2, { title, description });
+  const proposal = await proposeUpgrade(greeter.address, GreeterV2, { title, description, multisig });
 
   t.is(proposal.url, proposalUrl);
   sinon.assert.calledWithExactly(
@@ -48,7 +71,7 @@ test('proposes an upgrade', async t => {
       title,
       description,
       proxyAdmin: undefined,
-      via: undefined,
+      via: multisig,
       viaType: undefined,
     },
     {
@@ -66,7 +89,6 @@ test('proposes an upgrade with explicit multisig and proxy admin', async t => {
   const title = 'My upgrade';
   const description = 'My contract upgrade';
   const proxyAdmin = '0x20cE6FeEf8862CbCe65fd1cafA59ac8bbC77e445';
-  const multisig = '0xc0889725c22e2e36c524F41AECfddF5650432464';
   const multisigType = 'Gnosis Safe';
   const opts = { title, description, proxyAdmin, multisig, multisigType };
   const proposal = await proposeUpgrade(greeter.address, GreeterV2, opts);
@@ -95,7 +117,7 @@ test('proposes an upgrade reusing prepared implementation', async t => {
   fakeClient.proposeUpgrade.resolves({ url: proposalUrl });
 
   const greeterV2Impl = await upgrades.prepareUpgrade(greeter.address, GreeterV2);
-  const proposal = await proposeUpgrade(greeter.address, GreeterV2);
+  const proposal = await proposeUpgrade(greeter.address, GreeterV2, { multisig });
 
   t.is(proposal.url, proposalUrl);
   sinon.assert.calledWithExactly(
@@ -105,7 +127,7 @@ test('proposes an upgrade reusing prepared implementation', async t => {
       title: undefined,
       description: undefined,
       proxyAdmin: undefined,
-      via: undefined,
+      via: multisig,
       viaType: undefined,
     },
     {
@@ -132,4 +154,12 @@ test('fails if defender config is missing', async t => {
     message: 'Missing Defender API key and secret in hardhat config',
   });
   hre.config.defender = defender;
+});
+
+test('fails if multisig address is missing from UUPS proxy', async t => {
+  const { proposeUpgrade, fakeClient, greeter, GreeterV2 } = t.context;
+  sinon.assert.notCalled(fakeClient.proposeUpgrade);
+  await t.throwsAsync(() => proposeUpgrade(greeter.address, GreeterV2), {
+    message: 'Multisig address is a required property for UUPS proxies',
+  });
 });